+++ changelog.md
@@ -1,4 +1,7 @@
-# Change Log
-## 0.1.0 - Wed, 08 Dec 2021 16:09:09 +0700
-update script check reload rule: node conf and git files change
+# Change log
+
+- Mon, 28 Feb 2022 11:47:55 +0700
+Change push.py to push.sh using curl
+- Mon, 28 Feb 2022 11:58:31 +0700
+Add timeout on mkagent